Reserve some logic for future hardware: in prepare_kernel.sh

+++ user_space/prepare_kernel.sh
@@ -56,17 +56,17 @@ else
 fi
 
 if [ "$ARCH_OPTION" == "64" ]; then
-    LINUX_KERNEL_SRC_DIR_NAME=adi-linux-64
-    LINUX_KERNEL_CONFIG_FILE=$OPENWIFI_DIR/kernel_boot/kernel_config_zynqmp
-    ARCH_NAME="arm64"
-    CROSS_COMPILE_NAME="aarch64-linux-gnu-"
-    IMAGE_TYPE=Image
+  LINUX_KERNEL_SRC_DIR_NAME=adi-linux-64
+  LINUX_KERNEL_CONFIG_FILE=$OPENWIFI_DIR/kernel_boot/kernel_config_zynqmp
+  ARCH_NAME="arm64"
+  CROSS_COMPILE_NAME="aarch64-linux-gnu-"
+  IMAGE_TYPE=Image
 else
-    LINUX_KERNEL_SRC_DIR_NAME=adi-linux
-    LINUX_KERNEL_CONFIG_FILE=$OPENWIFI_DIR/kernel_boot/kernel_config
-    ARCH_NAME="arm"
-    CROSS_COMPILE_NAME="arm-linux-gnueabihf-"
-    IMAGE_TYPE=uImage
+  LINUX_KERNEL_SRC_DIR_NAME=adi-linux
+  LINUX_KERNEL_CONFIG_FILE=$OPENWIFI_DIR/kernel_boot/kernel_config
+  ARCH_NAME="arm"
+  CROSS_COMPILE_NAME="arm-linux-gnueabihf-"
+  IMAGE_TYPE=uImage
 fi
 
 home_dir=$(pwd)
@@ -80,37 +80,45 @@ git reset --hard
 cd $OPENWIFI_DIR/
 git submodule update $LINUX_KERNEL_SRC_DIR_NAME
 cd $OPENWIFI_DIR/$LINUX_KERNEL_SRC_DIR_NAME
-git fetch
-git checkout 2022_R2
-git pull origin 2022_R2
-# git reset --hard 2022_R2
-git reset --hard c2f371e014f0704be4db02e5014c51ae99477c13 # save this commit for tsn
+if false; then
+  echo "Reserve for future"
+else
+  git fetch
+  git checkout 2022_R2
+  git pull origin 2022_R2
+  # git reset --hard 2022_R2
+  git reset --hard c2f371e014f0704be4db02e5014c51ae99477c13 # save this commit for tsn
+fi
 
 source $XILINX_ENV_FILE
 export ARCH=$ARCH_NAME
 export CROSS_COMPILE=$CROSS_COMPILE_NAME
 
-# if [ "$ARCH_OPTION" == "64" ]; then
-  cp $LINUX_KERNEL_CONFIG_FILE ./.config
-  # cp $OPENWIFI_DIR/driver/ad9361/ad9361.c $OPENWIFI_DIR/$LINUX_KERNEL_SRC_DIR_NAME/drivers/iio/adc/ad9361.c -rf
-  # cp $OPENWIFI_DIR/driver/ad9361/ad9361_conv.c $OPENWIFI_DIR/$LINUX_KERNEL_SRC_DIR_NAME/drivers/iio/adc/ad9361_conv.c -rf
-  git apply ../kernel_boot/axi_hdmi_crtc.patch
-  git apply ../kernel_boot/ad9361.patch
-  git apply ../kernel_boot/ad9361_private.patch
-  git apply ../kernel_boot/ad9361_conv.patch
-  # #Ignore warning in mac80211 -- NOT necessary for 2022_R2 kernel!
-  # sed -i '3692 s/^/\/\//' ../$LINUX_KERNEL_SRC_DIR_NAME/net/mac80211/util.c
-# else
-  # make zynq_xcomm_adv7511_defconfig
-# fi
+if false; then
+  echo "Reserve for future"
+else
+  # if [ "$ARCH_OPTION" == "64" ]; then
+    cp $LINUX_KERNEL_CONFIG_FILE ./.config
+    # cp $OPENWIFI_DIR/driver/ad9361/ad9361.c $OPENWIFI_DIR/$LINUX_KERNEL_SRC_DIR_NAME/drivers/iio/adc/ad9361.c -rf
+    # cp $OPENWIFI_DIR/driver/ad9361/ad9361_conv.c $OPENWIFI_DIR/$LINUX_KERNEL_SRC_DIR_NAME/drivers/iio/adc/ad9361_conv.c -rf
+    git apply ../kernel_boot/axi_hdmi_crtc.patch
+    git apply ../kernel_boot/ad9361.patch
+    git apply ../kernel_boot/ad9361_private.patch
+    git apply ../kernel_boot/ad9361_conv.patch
+    # #Ignore warning in mac80211 -- NOT necessary for 2022_R2 kernel!
+    # sed -i '3692 s/^/\/\//' ../$LINUX_KERNEL_SRC_DIR_NAME/net/mac80211/util.c
+  # else
+    # make zynq_xcomm_adv7511_defconfig
+  # fi
 
-make oldconfig
-# make adi_zynqmp_defconfig
-make prepare && make modules_prepare
+  make oldconfig
+  # make adi_zynqmp_defconfig
+  make prepare && make modules_prepare
 
-# if [ "$#" -gt 2 ]; then
-make -j12 $IMAGE_TYPE UIMAGE_LOADADDR=0x8000
-make modules
-# fi
+  # if [ "$#" -gt 2 ]; then
+  make -j12 $IMAGE_TYPE UIMAGE_LOADADDR=0x8000
+  make modules
+  # fi
+fi
 
 cd $home_dir
